Add: system init config
System init config needs to read in the config file, flags, and env vars
but it would be different than the system config. The actual system
config needs to be lean, tidy, and well designed because many aspects of
it need to be shared and theres no real way around that.

This init config maps all the config fields to a struct and has a nice
unmarshal function that sets global instance. Because this is basically
just used for system basics like log level, it doesnt need have a mutex
because it wont be written to after `config.UnmarshalInitConfig`