Add: root cmd with init config logic
Add base cmd with config handling logic and nice logging. The init
config is for config file, env var, and flags. Because the actual shared
config between go routines is different and needs to come after tun
interface init, this has been broken up into init config and shared
config.

The logic for init config is initializing the logger. For a logger I
chose zerolog because its fast af and has a nice builder API which I
like. unfortunately the documentation is uhhh *weird*, it leaves out
some things.

For zerolog, I have set it up so the global logger is thread safe.
Because we will either have really pretty console logger which is thread
safe and not performant or writing to /var/log/buoy.log which is in
os.APPEND mode so the writes are thread safe, anway because of those our
global logger is thread safe.

This also holds the logic for setting log levels, where to write to,
unmarshalling init config, and reading config from /etc/buoy/buoy.yaml.

Also updating go version and adding dependancies like logging and cobra
and stuff.