Add: Shared config struct and functions
There needs to be a shared config of the actual running system. It has
to be thread safe, performant(enough), with nice errors. This adds most
of those.

It doesnt hold all needed fields like fleet peer list. It uses a RW
mutex to ensure that reading is async but writing is sync. This is
really a **WIP** but a general base is useful and needed