Add: initialization and creation of tun interface
A few things. So it turns out the actual address in a CIDR range is
important, guess I forgot that. Either way, to actually keep it, I
changed the config to unmarshal differently. I was running with the hook
to decode the CIDR and pull out the range but as I was saying, we need
the address too. I couldnt figure out a solution using mapstructure
cause Im lazy and its 2 AM.

The solution I picked works just fine. It unmarshals as a string and
then parses out afterwards.

The tun interface stuff is pretty cool ngl. Either way, instead of doing
it manually, like we were doing before, we dont do it manually. I wrote
in comments what each part does but basically it creates the interface
with a name, makes its range and address, brings it up, connects the tun
interface to it, and then returns the interface.

eventually we should make a route that destroys the interface.