Fix: ioctl invalid config tun,Rename: Init to New
Renamed Init to New. Bailey pointed out that naming it `Init` could be
confusing, `New` is much more unambiguous.

Fix the ioctl invalid config error that comes from `water.New`. This
error was caused by a few things so lets go over em.

1. This issue was caused by me creating a network interface with tenus
first. I thought that I had to create the interface and then configure
it and then create a TUN device over it. This was wrong. Changing the
order so water creates first fixes that.

2. Creating an interface with tenus would also cause an issue. Switching
to tenus setting a known interface also fixes that.