Add: translation package base & (de)compression
translation package base is done. It has a translate and untranslate
function. It also has unexported functions of encrypt/decrypt and
compression/decompression.

I also changed config to have password as a byte slice. This way we can
save time by casting once on init instead of casting every packet(1/10
time).

I chose gzip as a compression algorithm because it uses deflate on the
background but is more agressive on its actual compression. We need to
save size on the larger boys, I dont care about wasting CPU cycles on
the smaller packets.

I want to avoid IP fragmentation whenever possible so Im realllllllllly
betting on gzip to take care of it. There is still the *strong* chance
that its not enough and we should embrace fragmentation. Fuck
fragmentation, its a nightmare to implement ourselves.